third_party/WebKit/LayoutTests/crypto/wrapKey-badParameters-expected.txt

CONSOLE ERROR: Invalid key argument
CONSOLE ERROR: Invalid wrappingKey argument
CONSOLE ERROR: Algorithm: Not an object
CONSOLE ERROR: Invalid keyFormat argument
CONSOLE ERROR: Algorithm: SHA-1: Unsupported operation
CONSOLE ERROR: key.algorithm does not match that of operation
Tests calls to wrapKey() with bad inputs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
